مثال الطارة يعمل بنجاح، الحمد لله

+++ أمثلة/5.الدونت-الطارة/تجميع.js
@@ -9,18 +9,44 @@ const arabiMaps = require('@arabi/maps');
 
 const OUTPUT = path.resolve(__dirname, './donut.js');
 
-const map = arabiMaps.commonjs;
+const maps = arabiMaps.commonjs;
 
-map.global['عملية'] = ['process', {
-  خارج: ['stdout', { اكتب: 'write' }],
+maps.global['ضع_مؤقتا'] = 'setTimeout';
+
+maps.global['عملية'] = ['process', {
+  اخرج: 'exit',
+  المرة_القادمة: 'nextTick',
+  خارج: ['stdout', {
+    اكتب: 'write',
+    أعمدة: 'columns',
+    صفوف: 'rows',
+    احضر_حجم_النافذة: 'getWindowSize',
+    عندما: "on"
+  }],
 }];
 
+maps.global['افعل_فوريا'] = 'setImmediate';
+
+maps.modules = {
+  'هروب-الآنسي': ['ansi-escapes', {
+    'اخفي_المؤشر': 'cursorHide',
+    'اعد_المؤشر_لموضعه': 'cursorRestorePosition',
+    'احفظ_موضع_المؤشر': 'cursorSavePosition',
+    'اظهر_المؤشر': 'cursorShow',
+  }],
+
+  'حجم-النافذة': ['window-size', {
+    احضر: ['get', null, { returnMap: { اتساع: "width", ارتفاع: "height" } }]
+  }]
+}
+
 // delete the exsiting file
-fs.unlinkSync(OUTPUT);
+fs.existsSync(OUTPUT) && fs.unlinkSync(OUTPUT);
 
 arabi.translate({
   input: path.resolve(__dirname, './طارة.جس'),
   output: OUTPUT,
-  maps: map,
+  outputType: 'commonjs',
+  maps: maps,
   globalObject: "global"
 });
